Changed and renamed "dealHands" function to "deal" function.
the dealHands function is used once to deal hands at the start of the game. This function can be rewritten and simplified to deal x cards to players, with developers handling the use case. This allows the function to also be used when players receive cards from +2s or +4s, or draw cards themselves.

The function was also rewritten to use the "Player" class, and the main function, who was using the "dealHands" function was refactored accordingly.